Address PR feedback - add test to cover added logic
